Move utils folder into src
